requirements.txt
keras==2.10.0
numpy==1.23.3
paho_mqtt==1.6.1
pandas==1.5.0
requests==2.28.1
schedule==1.1.0
scikit_learn==1.1.2
